Appointment controller is updated

+++ src/controller/lib/appointment.controller.js
@@ -38,6 +38,9 @@ const createNewAppointment = asyncHandler(async (req, res) => {
 const getAllAppointments = asyncHandler(async (req, res) => {
   const uId = req.user._id;
   const appointment = await Appointment.find({ uId });
+  if (!appointment) {
+    throw new ApiError(400, "Appointments does not exist");
+  }
   res.status(200).json(new ApiResponse(200, appointment));
 });
 
@@ -45,6 +48,9 @@ const getPatientAllAppointments = asyncHandler(async (req, res) => {
   const pId = req.params?.id;
   if (!pId) throw new ApiError(400, "Provide patient id.");
   const appointment = await Appointment.find({ pId });
+  if (!appointment) {
+    throw new ApiError(400, "Appointments does not exist for this patient");
+  }
   res.status(200).json(new ApiResponse(200, appointment));
 });
 
@@ -52,12 +58,19 @@ const getAppointment = asyncHandler(async (req, res) => {
   const _id = req.params?.id;
   if (!_id) throw new ApiError(400, "Provide appointment id.");
   const appointment = await Appointment.findOne({ _id });
+  if (!appointment) {
+    throw new ApiError(400, "Appointment does not exist");
+  }
   res.status(200).json(new ApiResponse(200, appointment));
 });
 
 const AddPaymentShedule = asyncHandler(async (req, res) => {
   const apponintmentId = req.params?.id;
   if (!apponintmentId) throw new ApiError(400, "Provide patient id.");
+  const appointment = await Appointment.find({_id:apponintmentId})
+  if (!appointment) {
+    throw new ApiError(400, "Appointment does not exist");
+  }
   const newPaymentSchedule = await PaymentShcedule.create({
     apponintmentId,
     nCharges: req.body?.nCharges,
@@ -80,6 +93,9 @@ const deleteAppointment = asyncHandler(async (req, res) => {
   const _id = req.params?.id;
   if (!_id) throw new ApiError(400, "Provide appointment id.");
   const deletedAppointment = await Appointment.deleteOne({ _id });
+  if (!deletedAppointment?.acknowledged) {
+    throw new ApiError(400, "Appointment does not exist");
+  }
   await PaymentShcedule.deleteMany({ apponintmentId: deletedAppointment._id });
   res
     .status(200)
